Move classes to util module.
